Fix arange to avoid overflow.

+++ theano/tests/test_printing.py
@@ -342,7 +342,7 @@ def test_scan_debugprint2():
        | |Subtensor{int64} [@J] ''
        |   |Shape [@K] ''
        |   | |Subtensor{int64::} [@L] ''
-       |   |   |ARange{dtype='int16'} [@M] ''
+       |   |   |ARange{dtype='int64'} [@M] ''
        |   |   | |TensorConstant{0} [@N]
        |   |   | |TensorConstant{10000} [@O]
        |   |   | |TensorConstant{1} [@P]
@@ -366,7 +366,7 @@ def test_scan_debugprint2():
      > |coefficients[t] [@Y] -> [@S]
      > |Elemwise{pow,no_inplace} [@Z] ''
      >   |x_copy [@BA] -> [@W]
-     >   |<TensorType(int16, scalar)> [@BB] -> [@U]"""
+     >   |<TensorType(int64, scalar)> [@BB] -> [@U]"""
 
     for truth, out in zip(expected_output.split("\n"), lines):
         assert truth.strip() == out.strip()
@@ -425,7 +425,7 @@ def test_scan_debugprint3():
        | |Subtensor{int64} [@J] ''
        |   |Shape [@K] ''
        |   | |Subtensor{int64::} [@L] ''
-       |   |   |ARange{dtype='int8'} [@M] ''
+       |   |   |ARange{dtype='int64'} [@M] ''
        |   |   | |TensorConstant{0} [@N]
        |   |   | |TensorConstant{10} [@O]
        |   |   | |TensorConstant{1} [@P]
@@ -479,7 +479,7 @@ def test_scan_debugprint3():
      >   | | |Constant{1} [@BX]
      >   | |Constant{-1} [@BY]
      >   |DimShuffle{x} [@BZ] ''
-     >     |<TensorType(int8, scalar)> [@CA] -> [@U]
+     >     |<TensorType(int64, scalar)> [@CA] -> [@U]
 
     for{cpu,scan_fn} [@BE] ''
      >Elemwise{mul,no_inplace} [@CB] ''
